Move prctl_name() to util and fix its usage in code
