+++ prusa/link/printer_adapter/command_queue.py
@@ -11,7 +11,8 @@ from typing import Any, Dict, Optional
 from ..const import QUIT_INTERVAL
 from .command import Command, CommandFailed
 from .telemetry_passer import TelemetryPasser
-from .updatable import Thread, prctl_name
+from .updatable import Thread
+from ..util import prctl_name
 
 log = logging.getLogger(__name__)
